FastAPI/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
pydantic-settings==2.1.0
click==8.1.7
email-validator==2.1.0
python-multipart==0.0.20